Streamlit/01-Project-01-ToReadLocalFiles/requirement.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
python-docx
torch
transformers
regex
tqdm
nltk